feed-intel: update feed_wlan_6x hash to fix supplicants race
fix 1:
The teardown of the processes is managed from another location and
calling it from this location causes the wireless_process_kill_all
command to stuck and never return.

Removed the call to wireless_process_kill_all.

After the removal of the call to kill_all, we can use the config instead
of the data json file, since the file doesn't need to be recreated from
scratch.

fixt 2:
Due to racing, there is a case where the supplicant PID file is read
before the supplicant creates it. The result is that when the netifd
tries to close the supplicant - it doesn't hold the PID and thus doesn't
close it.

To overcome this, retry to read the PID file and validate exists before
continuing. This is done for up to 5 seconds which is more than enough
time for the supplicant to be up and PID file to be written.

PPM-1831.

Signed-off-by: Adam Dov <adov@maxlinear.com>